Deployment of Ceph cluster without controller nodes
This commit introduces a post-install hook that
adds a new release to fuel, in which there is no
lower limit on controller nodes, so it is possible
to deploy ceph cluster without having to deploy
openstack controller as well. ceph_release.json is
based on openstack.yaml, see generate_ceph_release.py
for more info.